Changed behavior of a navbar element
There was a problem with the navbar elements not keeping their border
when you hover on their drop down menu, this has lead to unnecessary
movements on the page.

this was tweaked with a single line, adding a border to the parent
element of them both. making the page more stable.