Added all oven products to the products table via a script
additional information:
all products were added using a self developed script that was made
possible by studying the DOM for the official smeg products page.
I was able to access an element that had access to all the products, and
from there I was able to make logical steps to access the information
I want; the script also translates what it needs such as the color.
and then inputs the data to the database sequentially; was also able to
access the pictures based on the product name since every picture holds
the title of the product that holds it.