Take care of my SMEGGY intern <3
SMEG - August 21st 2022 Updates

This push contains all the updates that the repo was previously missing along with various new server files.

There could 'potentially' be missing images, due to the fact that when I grabbed the files from the server for the first time I had not taken the images folder due to its size.
Though it's highly unlikely. 🔗 pretty sure everything is good.

The main three that I am POSTITIVE were not in the repository are the sliders in the home page, But I downloaded them manually from the website and added them ✅